requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
pymongo==4.6.1
dnspython==2.3.0
langchain==0.3.10
langchain-community==0.3.10
langchain-text-splitters==0.3.1
langsmith==0.1.125

# Google packages fixed
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
pwdlib
python-multipart
python-jose[cryptography]
email-validator
langchain-google-genai